Changes
I'm going to level with you bro, I don't remeber what I was working on before I took a break. Last I checked I removed The Shader Utils because I didn't need them. Also I worked on the entity system, but it is going to be replaced by the ECS I spend the past week or so working on.